Split ipc.ts into per-domain handler modules
Pure refactor — no behavior change. The 1591-line registerIpcHandlers
function is now an orchestrator that wires up domain-specific modules
under packages/desktop/src/main/handlers/.

- context.ts: IpcContext, AppState, AppContext factory with cached
  getters (config, dimensions, orgTree, accountMap, orgAccountsPath),
  withConnection, isCredentialError, toUserFriendlyError.
- query-utils.ts: result builders, queryAll, applyOrgTreeRollup,
  toNum/toStr/toEffort, isOwnerGroupBy, resolveEntityName.
- query.ts: query:* handlers (costs, daily-costs, trends, missing-tags,
  savings, entity-detail, filter-values).
- sync.ts: sync:status + data:* handlers + abort-controller registry.
- config.ts: config:* handlers (get, dimensions, org-tree).
- setup.ts: setup:* handlers (status, test-connection, list-profiles,
  list-buckets, browse-s3, write-config, scaffold-config).
- dimensions.ts: dimensions:* handlers (discover-tags, get/save-config).
- savings.ts, ui.ts: per-feature preferences handlers.
- org.ts: AWS Organizations sync handlers + progress state.
- auto-sync.ts: auto-sync:* handlers + bootstrap on launch.

ipc.ts shrinks from 1591 lines to 25.
